requirements.txt
Click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dotenv==0.16.0
SQLAlchemy==1.4.3
Werkzeug==1.0.1
pytest==6.2.2
requests~=2.25.1
python-dateutil~=2.8.1